Refactor: Move the pipeline classes to a separate folder
There's only two right now, but at least three more or so will be needed shortly.
